refactoring: move method findTalksByConferenceId() from ConferenceService to TalkService <eom>
